Potential fix for code scanning alert no. 375: Result of multiplication cast to wider type

+++ src/test/java/co/rsk/federate/FederationProviderFromFederatorSupportTest.java
@@ -184,7 +184,7 @@ class FederationProviderFromFederatorSupportTest {
     private void mockFederationMemberKeys(int i) {
         when(federatorSupportMock.getFederatorPublicKeyOfType(i, FederationMember.KeyType.BTC)).thenReturn(ECKey.fromPrivate(BigInteger.valueOf((i +1)*1000)));
         when(federatorSupportMock.getFederatorPublicKeyOfType(i, FederationMember.KeyType.RSK)).thenReturn(ECKey.fromPrivate(BigInteger.valueOf((i +1)*1000+1)));
-        when(federatorSupportMock.getFederatorPublicKeyOfType(i, FederationMember.KeyType.MST)).thenReturn(ECKey.fromPrivate(BigInteger.valueOf((i +1)*1000+2)));
+        when(federatorSupportMock.getFederatorPublicKeyOfType(i, FederationMember.KeyType.MST)).thenReturn(ECKey.fromPrivate(BigInteger.valueOf(((long) (i + 1)) * 1000 + 2)));
     }
 
     @Test
